Put together the first stable version of the app
Arranged the app elements using a combination of Flexbox, for
alignment and one-directional layout purposes, and CSS Grid, which came
in handy especially to sort the gameboard out.

The JS followed design patterns such as the module and the function
factory ones, so as not to pollute the window object. The script itself
was all about tying the gameboard spaces to the DOM and replicating
the logic behind the win/tie conditions.

The color palette ranged from lime to teal, and was picked by hand
from the Tailwind Docs. There were no major assets aside from a play
icon from Material Design Icons.